[#128] Proper multi-module maven project structure * Extract the custom JSP tag implementations to a separate module
