+++ csrfguard-extensions/csrfguard-jsp-tags/src/main/java/org/owasp/csrfguard/tag/TokenValueTag.java
@@ -33,24 +33,24 @@ import org.owasp.csrfguard.CsrfGuard;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
+import java.util.Objects;
 
-public final class TokenTag extends AbstractUriTag {
+public final class TokenValueTag extends AbstractUriTag {
 
-	private final static long serialVersionUID = 0x12164baa;
+	private final static long serialVersionUID = 0xaaca46d3;
 
 	@Override
 	public int doStartTag() {
 		final CsrfGuard csrfGuard = CsrfGuard.getInstance();
-		final String tokenName = csrfGuard.getTokenName();
 
-		if (csrfGuard.isTokenPerPageEnabled() && getUri() == null) {
-			throw new IllegalStateException("must define 'uri' attribute when token per page is enabled");
+		if (csrfGuard.isTokenPerPageEnabled() && Objects.isNull(getUri())) {
+			throw new IllegalStateException("Must define 'uri' attribute when token per page is enabled");
 		}
 
 		final String tokenValue = csrfGuard.getTokenService().getTokenValue((HttpServletRequest) this.pageContext.getRequest(), getUri());
 
 		try {
-			this.pageContext.getOut().write(tokenName + '=' + tokenValue);
+			this.pageContext.getOut().write(tokenValue);
 		} catch (final IOException e) {
 			this.pageContext.getServletContext().log(e.getLocalizedMessage(), e);
 		}
